Move certain syntax generator internal methods down to shared layer
