+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Formatting/FormattingExtensions.cs
@@ -279,5 +279,72 @@ namespace Microsoft.CodeAnalysis.Formatting
 
             return new TextChange(span, newText);
         }
+
+        internal static IEnumerable<TextSpan> GetAnnotatedSpans(SyntaxNode node, SyntaxAnnotation annotation)
+        {
+            foreach (var nodeOrToken in node.GetAnnotatedNodesAndTokens(annotation))
+            {
+                var firstToken = nodeOrToken.IsNode ? nodeOrToken.AsNode().GetFirstToken(includeZeroWidth: true) : nodeOrToken.AsToken();
+                var lastToken = nodeOrToken.IsNode ? nodeOrToken.AsNode().GetLastToken(includeZeroWidth: true) : nodeOrToken.AsToken();
+                yield return GetSpan(firstToken, lastToken);
+            }
+        }
+
+        internal static TextSpan GetSpan(SyntaxToken firstToken, SyntaxToken lastToken)
+        {
+            var previousToken = firstToken.GetPreviousToken();
+            var nextToken = lastToken.GetNextToken();
+
+            if (previousToken.RawKind != 0)
+            {
+                firstToken = previousToken;
+            }
+
+            if (nextToken.RawKind != 0)
+            {
+                lastToken = nextToken;
+            }
+
+            return TextSpan.FromBounds(firstToken.SpanStart, lastToken.Span.End);
+        }
+
+        internal static IEnumerable<TextSpan> GetElasticSpans(SyntaxNode root)
+        {
+            var tokens = root.GetAnnotatedTrivia(SyntaxAnnotation.ElasticAnnotation).Select(tr => tr.Token).Distinct();
+            return AggregateSpans(tokens.Select(t => GetElasticSpan(t)));
+        }
+
+        internal static TextSpan GetElasticSpan(SyntaxToken token)
+            => GetSpan(token, token);
+
+        private static IEnumerable<TextSpan> AggregateSpans(IEnumerable<TextSpan> spans)
+        {
+            var aggregateSpans = new List<TextSpan>();
+
+            var last = default(TextSpan);
+            foreach (var span in spans)
+            {
+                if (last == default)
+                {
+                    last = span;
+                }
+                else if (span.IntersectsWith(last))
+                {
+                    last = TextSpan.FromBounds(last.Start, span.End);
+                }
+                else
+                {
+                    aggregateSpans.Add(last);
+                    last = span;
+                }
+            }
+
+            if (last != default)
+            {
+                aggregateSpans.Add(last);
+            }
+
+            return aggregateSpans;
+        }
     }
 }
